Boilerplate, Custom Attribute
Creating custom attribute, need to really analyze the requirements of it and how extensive the code should be, not very secure nor modular, but doesn't need to be very modular as it's for one class for now.

Should probs just go full boilerplate and fully implement a custom attribute with custom modifiers so as to keep this scaleable for future ideas, which are currently none.

Hence the needed analysis