Rat Prototype
First Entity consisting of the Rat, currently the translation key is bugged referring to it as a pig for some reason, been hunting down why but so far no clue. Model currently clips into the floor a little, will edit. Rat is a passive creature that will attempt to flee any players and is quite quick, it uses Bat noises for squeaks.